Move ObjectDetectionImageSegmentation to own file
